add: regionとareaを取得するAPIを作成
