+++ routes/area/areaService.ts
@@ -1,24 +1,42 @@
-import { Response, Request, Router } from 'express';
+import { PrismaClient } from '@prisma/client';
 
-import { getArea, getAreaDetail } from './areaControl';
-const router = Router();
+const prisma = new PrismaClient();
 
-/**
- * 地区一覧を取得する
- */
-router.get('/', async (req: Request, res: Response): Promise<void> => {
-  const area = await getArea();
-  res.json(area);
-});
+export const getArea = async () => {
+  const area = await prisma.area.findMany({
+    select: {
+      id: true,
+      area_name: true,
+    },
+  });
 
-/**
- * 指定されたidの地区名とレビューを返す
- *
- */
-router.get('/:id', async (req: Request, res: Response): Promise<void> => {
-  const { id } = req.params;
-  const [areaName, areaDetails, category] = await getAreaDetail(id);
-  res.json({ name: areaName, areaDetails, category });
-});
+  return area;
+};
 
-export const area = router;
+export const getAreaDetail = async (id: string) => {
+  const area = await prisma.area.findMany({
+    where: {
+      id: Number(id), // Numberにキャスト
+    },
+    select: {
+      area_name: true,
+    },
+  });
+  const areaName = area[0].area_name;
+  const areaDetails = await prisma.area
+    .findUnique({
+      where: {
+        id: Number(id),
+      },
+    })
+    .reviews(); // areaのreviews[]をリレーションして取得
+
+  const category = await prisma.category.findMany({
+    select: {
+      id: true,
+      category_name: true,
+    },
+  });
+
+  return [areaName, areaDetails, category];
+};
